Function github_rest::methods::comment_on_commit

pub async fn comment_on_commit<T>(
    client: &T, 
    owner: impl Into<String>, 
    repo: impl Into<String>, 
    sha: impl Into<String>, 
    options: &CommitCommentBody
) -> Result<CommitComment, GithubRestError> where
    T: Requester, 

tags repos

post /repos/{owner}/{repo}/commits/{commit_sha}/comments

docs https://docs.github.com/rest/reference/repos#create-a-commit-comment

Create a commit comment Create a comment for a commit using its :commit_sha.

This endpoint triggers notifications. Creating content too quickly using this endpoint may result in secondary rate limiting. See “Secondary rate limits” and “Dealing with secondary rate limits” for details.