Function github_rest::methods::get_commits

pub async fn get_commits<T>(
    client: &T,
    owner: impl Into<String>,
    repo: impl Into<String>,
    options: Option<&GetCommitsBody>
) -> Result<Commits, GithubRestError> where
    T: Requester, 

tags repos

get /repos/{owner}/{repo}/commits

docs https://docs.github.com/rest/reference/repos#list-commits

List commits Signature verification object

The response will include a verification object that describes the result of verifying the commit’s signature. The following fields are included in the verification object:

Name	Type	Description

verified	boolean	Indicates whether GitHub considers the signature

in this commit to be verified.		reason

verified value. Possible values and their meanings are enumerated in table

below.		signature

the commit.		payload

These are the possible values for reason in the verification object:

Value	Description

expired_key	The key that made the signature is expired.

not_signing_key	The “signing” flag is not among the usage flags in the

GPG key that made the signature.

communicating with the signature verification service.

gpgverify_unavailable	The signature verification service is currently

unavailable.

unknown_signature_type	A non-PGP signature was found in the commit.

no_user	No user was associated with the committer email address in

the commit.

commit was associated with a user, but the email address is not verified on

her/his account.

commit is not included in the identities of the PGP key that made the

signature.

registered with any user’s account.

error parsing the signature.

cryptographically verified using the key whose key-id was found in the

signature.

is considered to be verified.