Function github_rest::methods::get_pulls

pub async fn get_pulls<T>(
    client: &T, 
    owner: impl Into<String>, 
    repo: impl Into<String>, 
    options: Option<&GetPullsBody>
) -> Result<Pulls, GithubRestError> where
    T: Requester, 

tags pulls

get /repos/{owner}/{repo}/pulls

docs https://docs.github.com/rest/reference/pulls#list-pull-requests

List pull requests Draft pull requests are available in public repositories with GitHub Free and GitHub Free for organizations, GitHub Pro, and legacy per-repository billing plans, and in public and private repositories with GitHub Team and GitHub Enterprise Cloud. For more information, see GitHub’s products in the GitHub Help documentation.