Module github_rest::model::event_types

Structs

OrgEventInfo

Used to represent the base fields provided by events originating from organizations.

RepoEventInfo

Used to represent the base fields provided by events originating from repositories.

Enums

EventTypes

Used to represent all possible values for the x-github-event header sent with all webhook payloads.

Traits

Event

The base trait used to represent different types of events. This will eventually have some subtraits with convenience methods.