Trait github_rest::model::user::GitHubUser

pub trait GitHubUser {
    fn add_to_org<'life0, 'life1, 'async_trait, T, A>(
        &'life0 self, 
        client: &'life1 T, 
        org: A, 
        role: Option<Role>
    ) -> Pin<Box<dyn Future<Output = Result<AddToOrgResponse, GithubRestError>> + Send + 'async_trait>>
    where
        T: Requester,
        A: Into<String> + Send,
        T: 'async_trait,
        A: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_following<'life0, 'life1, 'async_trait, T>(
        &'life0 self, 
        client: &'life1 T, 
        followers_per_page: Option<u8>, 
        page_number: Option<u8>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SimpleUser>, GithubRestError>> + Send + 'async_trait>>
    where
        T: Requester,
        T: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn get_followers<'life0, 'async_trait, T>(
        self, 
        client: &'life0 T, 
        followers_per_page: Option<u8>, 
        page_number: Option<u8>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<SimpleUser>, GithubRestError>> + Send + 'async_trait>>
    where
        T: Requester,
        T: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;
}

A trait for all GitHub user structs. Intended for SimpleUser, PublicUser, and PrivateUser

Required Methods

fn add_to_org<'life0, 'life1, 'async_trait, T, A>(
&'life0 self,
client: &'life1 T,
org: A,
role: Option<Role>
) -> Pin<Box<dyn Future<Output = Result<AddToOrgResponse, GithubRestError>> + Send + 'async_trait>> where
T: Requester,
A: Into<String> + Send,
T: 'async_trait,
A: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

fn get_following<'life0, 'life1, 'async_trait, T>(
&'life0 self,
client: &'life1 T,
followers_per_page: Option<u8>,
page_number: Option<u8>
) -> Pin<Box<dyn Future<Output = Result<Vec<SimpleUser>, GithubRestError>> + Send + 'async_trait>> where
T: Requester,
T: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Get a list of the users that the user in question is following.

followers_per_page - The number of users to get per page. Default is 30.

page_number - The page number of the result to return. Default is 1.

fn get_followers<'life0, 'async_trait, T>(
self,
client: &'life0 T,
followers_per_page: Option<u8>,
page_number: Option<u8>
) -> Pin<Box<dyn Future<Output = Result<Vec<SimpleUser>, GithubRestError>> + Send + 'async_trait>> where
T: Requester,
T: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

Get a list of the users that are following the user in question.

followers_per_page - The number of followers to get per page. Default is 30.

page_number - The page number of the result to return. Default is 1.