Function github_rest::methods::get_issues

pub async fn get_issues<T>(
    client: &T,
    owner: impl Into<String>,
    repo: impl Into<String>,
    options: Option<&GetIssuesBody>
) -> Result<Issues, GithubRestError>where
    T: Requester,

tags issues

get /repos/{owner}/{repo}/issues

docs https://docs.github.com/rest/reference/issues#list-repository-issues

List repository issues List issues in a repository.

Note: GitHub’s REST API v3 considers every pull request an issue, but not every issue is a pull request. For this reason, “Issues” endpoints may return both issues and pull requests in the response. You can identify pull requests by the pull_request key. Be aware that the id of a pull request returned from “Issues” endpoints will be an issue id. To find out the pull request id, use the “List pull requests” endpoint.