Trait github_rest::model::event_types::Event

pub trait Event<'de>: Serialize + Deserialize<'de> {
    type Origin: Serialize + Deserialize<'de>;
    fn origin(&self) -> Option<Self::Origin> { ... }
}

The base trait used to represent different types of events. This will eventually have some subtraits with convenience methods.

All event types implement this.

Associated Types

type Origin: Serialize + Deserialize<'de>

Provided methods

fn origin(&self) -> Option<Self::Origin>